requirements.txt
aiohttp
requests
pandas
websocket-client
